% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_median_value_categorical.R
\name{make_median_value_categorical}
\alias{make_median_value_categorical}
\title{Test}
\usage{
make_median_value_categorical(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Test
}
