% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefile_map.R
\name{make_shapefile_map}
\alias{make_shapefile_map}
\title{Title}
\usage{
make_shapefile_map(df, tooltip_column, pre_tooltip_text)
}
\arguments{
\item{pre_tooltip_text}{}
}
\value{

}
\description{
Title
}
